+++ mitra_workers/src/periodic_tasks.rs
@@ -32,14 +32,14 @@ use mitra_models::{
 };
 use mitra_utils::datetime::days_before_now;
 
-use crate::payments::{
-    common::update_expired_subscriptions,
-    monero::{check_closed_invoices, check_monero_subscriptions},
-};
 use super::importer::{
     import_followers_task,
     import_follows_task,
     ImporterJobData,
+};
+use super::payments::{
+    common::update_expired_subscriptions,
+    monero::{check_closed_invoices, check_monero_subscriptions},
 };
 
 pub async fn subscription_expiration_monitor(